Problems 2 & 7 solved!
Problem 3 needs considerably more thought but was uploaded as well. Currently the algo I'm using to acquire the answer is slow and quite clunky. I don't mean for this to be the case and hope to be able to sort it out soon. I know there's a more efficient way of solving this simply because all Euler problems can be solved within a minute apparently (but writing the right code can take days they say so you'll need to work for it!)